Extracted generic IsAuthorized into Filter.cs so it's not comingled with controller specific logic.
