+++ src/aadcapi/App_Start/ControllerAuthorizationConfig.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using aadcapi.Models;
+using aadcapi.Utils.Authorization;
 
 namespace aadcapi
 {
